requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiohttp-cors==0.7.0
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
blessed==1.20.0
blinker==1.8.2
brax
cachetools==5.3.3
certifi==2024.6.2
cffi==1.17.0
charset-normalizer==3.3.2
chex==0.1.84
click==8.1.7
cloudpickle==3.0.0
colorful==0.5.6
contextlib2==21.6.0
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.11
-e git+https://github.com/Farama-Foundation/D4RL.git@71a9549f2091accff93eeff68f1f3ab2c0e0a288#egg=D4RL
decorator==5.1.1
distlib==0.3.8
distrax==0.1.5
dm-env==1.6
dm-tree==0.1.8
dm_control==1.0.21
docker-pycreds==0.4.0
dotmap==1.3.30
etils==1.7.0
evosax==0.1.5
exceptiongroup==1.2.1
executing==2.0.1
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.15.4
flashbax==0.1.0
Flask==3.0.3
Flask-Cors==4.0.1
flax
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
glfw==2.7.0
google-api-core==2.19.1
google-auth==2.31.0
googleapis-common-protos==1.63.2
gpustat==1.1.1
grpcio==1.64.1
gym
gym-notices==0.0.8
gymnasium==0.29.1
gymnax
h5py==3.11.0
hydra-core==1.3.2
idna==3.7
imageio==2.34.2
importlib_resources==6.4.0
iniconfig==2.0.0
ipython==8.26.0
itsdangerous==2.2.0
jax==0.4.25
# jaxlib==0.4.25+cuda12.cudnn89
jaxopt==0.8.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
labmaze==1.0.6
linkify-it-py==2.0.3
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
mediapy==1.2.2
memray==1.13.3
# mjrl==1.0.0
ml-dtypes==0.4.0
ml_collections==0.1.1
msgpack==1.0.8
mujoco==3.2.1
mujoco-mjx==3.1.3
mujoco-py==2.1.2.14
multidict==6.0.5
nest-asyncio==1.6.0
numpy==1.26.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu12==12.4.131
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu12==11.2.1.3
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.550.52
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
omegaconf==2.3.0
opencensus==0.11.4
opencensus-context==0.1.3
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.5.16
packaging==24.1
pandas==2.2.2
parso==0.8.4
pettingzoo==1.24.3
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-spy==0.3.14
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybullet==3.2.5
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.6.0
Pygments==2.18.0
PyOpenGL==3.1.7
pyparsing==3.1.2
pytest==8.2.2
python-dateutil==2.9.0.post0
pytinyrenderer==0.0.14
pytz==2024.1
PyYAML==6.0.1
ray==2.31.0
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.2
scipy==1.10.1
sentry-sdk==2.7.1
setproctitle==1.3.3
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
stack-data==0.6.3
tensorboardX==2.6.2.2
tensorflow-probability==0.24.0
tensorstore==0.1.63
termcolor==2.4.0
textual==0.71.0
tomli==2.0.1
toolz==0.12.1
tqdm==4.66.4
traitlets==5.14.3
trimesh==4.4.1
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.2
virtualenv==20.26.3
wandb==0.17.4
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
xvfbwrapper==0.2.9
yarl==1.9.4
zipp==3.19.2
